solving the server error, changing to SQLite from Postgress

+++ backend/app/models.py
@@ -16,14 +16,17 @@ class User(db.Model):
     def check_password(self, password):
         return check_password_hash(self.password_hash, password)
 
+
 class Flower(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100), nullable=False)
     description = db.Column(db.Text)
-    price = db.Column(db.Numeric(10,2), nullable=False)
+    # Use Float for SQLite compatibility (Numeric can behave differently)
+    price = db.Column(db.Float, nullable=False)
     image_url = db.Column(db.Text)
     florist_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     created_at = db.Column(db.DateTime, default=datetime.utcnow)
+
 
 class Order(db.Model):
     id = db.Column(db.Integer, primary_key=True)
@@ -31,11 +34,12 @@ class Order(db.Model):
     florist_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     flower_id = db.Column(db.Integer, db.ForeignKey('flower.id'))
     quantity = db.Column(db.Integer, nullable=False)
-    total_price = db.Column(db.Numeric(10,2), nullable=False)
+    total_price = db.Column(db.Float, nullable=False)  # Changed from Numeric to Float
     status = db.Column(db.String(50), default='pending')
-    delivery_lat = db.Column(db.Numeric(9,6))
-    delivery_lng = db.Column(db.Numeric(9,6))
+    delivery_lat = db.Column(db.Float)  # Changed from Numeric
+    delivery_lng = db.Column(db.Float)  # Changed from Numeric
     created_at = db.Column(db.DateTime, default=datetime.utcnow)
+
 
 class Message(db.Model):
     id = db.Column(db.Integer, primary_key=True)
